+++ test/ReaderTaskEither.ts
@@ -281,15 +281,11 @@ describe('ReaderTaskEither', () => {
   it('getApplicativeReaderTaskValidation', async () => {
     const A = _.getApplicativeReaderTaskValidation(T.ApplicativePar, semigroupString)
     assert.deepStrictEqual(await sequenceT(A)(_.left('a'), _.left('b'))(null)(), E.left('ab'))
-    const AV = _.getReaderTaskValidation(semigroupString)
-    assert.deepStrictEqual(await sequenceT(AV)(_.left('a'), _.left('b'))(null)(), E.left('ab'))
   })
 
   it('getAltReaderTaskValidation', async () => {
     const A = _.getAltReaderTaskValidation(semigroupString)
     assert.deepStrictEqual(await A.alt(_.left('a'), () => _.left('b'))(null)(), E.left('ab'))
-    const AV = _.getReaderTaskValidation(semigroupString)
-    assert.deepStrictEqual(await AV.alt(_.left('a'), () => _.left('b'))(null)(), E.left('ab'))
   })
 
   describe('bracket', () => {