+++ CHANGELOG.md
@@ -54,6 +54,7 @@ high state of flux, you're at risk of it changing without notice.
     - remove `run` (@gcanti)
   - `ReaderTaskEither`
     - remove `local`, use `Reader`'s `local` instead (@gcanti)
+    - remove `getReaderTaskValidation`, use `getApplicativeReaderTaskValidation` and `getAltReaderTaskValidation` instead (@gcanti)
 
 # 2.9.1